Add pointer to generic shell
