+++ opencog/cogserver/server/Module.h
@@ -12,8 +12,6 @@
  * from the module author.
  */
 
-#include <string>
-
 #ifndef _OPENCOG_MODULE_H
 #define _OPENCOG_MODULE_H
 
@@ -45,36 +43,6 @@ namespace opencog
     inline const char * MODNAME::id(void) {                           \
         return "opencog::" #MODNAME;                                  \
     }
-
-
-/**
- * DEFINE_SHELL_MODULE -- Declare a new C++ Module class, suitable
- * for providing a network shell. Network shells can respond to inputs
- * coming over a network TCP/IP socket. All the details of socket
- * handling are abstracted away.
- */
-#define DEFINE_SHELL_MODULE(MODNAME)                                  \
-namespace opencog {                                                   \
-                                                                      \
-class MODNAME : public Module {                                       \
-    private:                                                          \
-        class shelloutRequest : public Request                        \
-        {                                                             \
-            public:                                                   \
-                static const RequestClassInfo& info(void);            \
-                shelloutRequest(CogServer& cs) : Request(cs) {};      \
-                virtual ~shelloutRequest() {};                        \
-                virtual bool execute(void);                           \
-                virtual bool isShell(void) { return true; }           \
-        };                                                            \
-        Factory<shelloutRequest, Request> shelloutFactory;            \
-    public:                                                           \
-        MODNAME(CogServer&);                                          \
-        virtual ~MODNAME();                                           \
-        static const char *id(void);                                  \
-        virtual void init(void);                                      \
-        virtual bool config(const char *);                            \
-}; }
 
 /**
  * This class defines the base abstract class that should be extended
