+++ opencog/cogserver/shell/SexprShellModule.cc
@@ -28,6 +28,7 @@
 #include <opencog/cogserver/server/ServerConsole.h>
 
 #include "SexprShell.h"
+#include "ShellModule.h"
 
 using namespace opencog;
